% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_extract_helpers.R
\name{.resultColumns}
\alias{.resultColumns}
\title{Return column names to be used for summary operations}
\usage{
.resultColumns(
  value_names,
  weight_names,
  fun,
  full_colnames,
  quantiles = numeric(),
  unique_values = numeric(),
  colname_fun = NULL
)
}
\arguments{
\item{value_names}{names of value raster layers}

\item{weight_names}{names of weighting raster layers}

\item{fun}{functions or names of summary operations}

\item{full_colnames}{return a complete column name even when there is no
ambiguity?}

\item{quantiles}{quantiles to use when \code{stat_names} contains \code{quantile}}
}
\value{
character vector of column names
}
\description{
Return column names to be used for summary operations
}
\keyword{internal}
