% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_extract_helpers.R
\name{.valueWeightIndexes}
\alias{.valueWeightIndexes}
\title{Compute indexes for the value and weight layers that should be
processed together}
\usage{
.valueWeightIndexes(num_values, num_weights)
}
\arguments{
\item{num_values}{number of layers in value raster}

\item{num_weights}{number of layers in weighting raster}
}
\value{
list with \code{values} and \code{weights} elements
        providing layer indexes
}
\description{
Compute indexes for the value and weight layers that should be
processed together
}
\keyword{internal}
