% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactextractr-package.R
\docType{package}
\name{exactextractr-package}
\alias{exactextractr}
\alias{exactextractr-package}
\title{exactextractr: Fast Extraction from Raster Datasets using Polygons}
\description{
Quickly and accurately summarizes raster values over polygonal areas ("zonal statistics").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://isciences.gitlab.io/exactextractr/}
  \item \url{https://github.com/isciences/exactextractr}
  \item Report bugs at \url{https://github.com/isciences/exactextractr/issues}
}

}
\author{
\strong{Maintainer}: Daniel Baston \email{dbaston@isciences.com}

Other contributors:
\itemize{
  \item ISciences, LLC [copyright holder]
}

}
\keyword{internal}
