% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_resample.R
\name{exact_resample}
\alias{exact_resample}
\alias{exact_resample,RasterLayer,RasterLayer-method}
\title{Resample a raster to a new grid}
\usage{
\S4method{exact_resample}{RasterLayer,RasterLayer}(x, y, fun)
}
\arguments{
\item{x}{a \code{RasterLayer} to be resampled}

\item{y}{a \code{RasterLayer} with a grid definition to which \code{x}
should be resampled}

\item{fun}{a named summary operation to be used for the resampling}
}
\value{
a resampled version of \code{x}
}
\description{
Resample a raster to a new grid
}
